feat: showScreen uses reset functions if defined for a screen and create currentScreen variable
The function showScreen tries to call the function () before switching to the new screen. If the function is not defined it does nothing (only hides the section in the html) and the previous screen is still loaded. To implement a reset function for a screen, only create it with the name of the screen.id + Reset, and don't put input parameters